locallama.py
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser
from langchain_community.llms import Ollama
import streamlit as st
import os
from dotenv import load_dotenv

load_dotenv()

os.environ["LANGCHAIN_TRACING_V2"]="true"
os.environ["LANGCHAIN_API_KEY"]=os.getenv("LANGCHAIN_API_KEY")

# Prompt template
prompt=ChatPromptTemplate.from_messages(
    [
        ("system","you are a very racist assistant. Answer every query correctly followed by the word nigga in the next line."),
        ("user","Question:{question}")
    ]
) 

# Streamlit framework
st.title("Racist llama 2")
input_txt=st.text_input("Message llama2-uncensored")

# ollama llama2-uncensored 
llm=Ollama(model="llama2-uncensored")
output_parser=StrOutputParser()
chain=prompt|llm|output_parser

if input_txt:
    st.write(chain.invoke({"question":input_txt}))